+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 - icons and images to landing page
 
 ### Changed
+- timed display route to carousel
 - Departure station can be changed
 - category icons to filled and line icons
 - Answer animation can be skipped